job_screening_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
PyMuPDF==1.24.4
pandas==2.1.1
nltk==3.8.1
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
email-validator==2.0.0